# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@daytona.com
    name: Daytona Platforms Inc.
    url: https://www.daytona.io
  description: Daytona AI platform API Docs
  title: Daytona
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
...
  /workspace:
    get:
      operationId: listWorkspaces
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Include verbose output
          explode: true
          in: query
          name: verbose
          required: false
          schema:
            type: boolean
          style: form
        - description: JSON encoded labels to filter by
          explode: true
          in: query
          name: labels
          required: false
          schema:
            example: '{"label1": "value1", "label2": "value2"}'
            type: string
          style: form
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Workspace'
                type: array
          description: List of all workspaces
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: List all workspaces
      tags:
        - workspace
    post:
      operationId: createWorkspace
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspace'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: The workspace has been successfully created.
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create a new workspace
      tags:
        - workspace
  /workspace/{workspaceId}:
    delete:
      operationId: deleteWorkspace
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
        - explode: true
          in: query
          name: force
          required: true
          schema:
            type: boolean
          style: form
      responses:
        '200':
          description: Workspace has been deleted
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Delete workspace
      tags:
        - workspace
    get:
      operationId: getWorkspace
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Include verbose output
          explode: true
          in: query
          name: verbose
          required: false
          schema:
            type: boolean
          style: form
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: Workspace details
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get workspace details
      tags:
        - workspace
  /workspace/{workspaceId}/start:
    post:
      operationId: startWorkspace
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Workspace has been started
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Start workspace
      tags:
        - workspace
  /workspace/{workspaceId}/stop:
    post:
      operationId: stopWorkspace
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Workspace has been stopped
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Stop workspace
      tags:
        - workspace
  /workspace/{workspaceId}/labels:
    put:
      operationId: replaceLabels
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceLabels'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceLabels'
          description: Labels have been successfully replaced
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Replace workspace labels
      tags:
        - workspace
  /workspace/{workspaceId}/snapshot:
    post:
      operationId: createSnapshot
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
          description: Workspace snapshot has been initiated
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Create workspace snapshot
      tags:
        - workspace
  /workspace/{workspaceId}/public/{isPublic}:
    post:
      operationId: updatePublicStatus
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
        - description: Public status to set
          explode: false
          in: path
          name: isPublic
          required: true
          schema:
            type: boolean
          style: simple
      responses:
        '201':
          description: ''
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Update public status
      tags:
        - workspace
  /workspace/{workspaceId}/autostop/{interval}:
    post:
      operationId: setAutostopInterval
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
        - description: Auto-stop interval in minutes (0 to disable)
          explode: false
          in: path
          name: interval
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          description: Auto-stop interval has been set
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Set workspace auto-stop interval
      tags:
        - workspace
  /workspace/{workspaceId}/archive:
    post:
      operationId: archiveWorkspace
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Workspace has been archived
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Archive workspace
      tags:
        - workspace
  /workspace/{workspaceId}/ports/{port}/preview-url:
    get:
      operationId: getPortPreviewUrl
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
        - description: Port number to get preview URL for
          explode: false
          in: path
          name: port
          required: true
          schema:
            type: number
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortPreviewUrl'
          description: Preview URL for the specified port
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get preview URL for a workspace port
      tags:
        - workspace
  /workspace/{workspaceId}/build-logs:
    get:
      operationId: getBuildLogs
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Build logs stream
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getBuildLogs_400_response'
          description: Bad request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getBuildLogs_400_response'
          description: Build logs not found
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get build logs
      tags:
        - workspace
...
components:
  schemas:
...
    WorkspaceInfo:
      properties:
        created:
          description: The creation timestamp of the project
          example: 2023-10-01T12:00:00Z
          type: string
        name:
          default: ''
          deprecated: true
          description: 'Deprecated: The name of the workspace'
          example: MyWorkspace
          type: string
        providerMetadata:
          description: Additional metadata provided by the provider
          example: '{"key": "value"}'
          type: string
      required:
        - created
        - name
      type: object
...
    Workspace:
      example:
        image: daytonaio/workspace:latest
        memory: 4
        volumes:
          - mountPath: /data
            volumeId: volume123
          - mountPath: /data
            volumeId: volume123
        cpu: 2
        env:
          NODE_ENV: production
        snapshotCreatedAt: 2024-10-01T12:00:00Z
        gpu: 0
        autoStopInterval: 30
        labels:
          daytona.io/public: 'true'
        target: local
        organizationId: organization123
        disk: 10
        snapshotState: None
        public: false
        errorReason: The workspace is not running
        name: MyWorkspace
        id: workspace123
        state: creating
        user: daytona
        info: ''
      properties:
        id:
          description: The ID of the workspace
          example: workspace123
          type: string
        name:
          default: ''
          deprecated: true
          description: The name of the workspace
          example: MyWorkspace
          type: string
        organizationId:
          description: The organization ID of the workspace
          example: organization123
          type: string
        image:
          description: The image used for the workspace
          example: daytonaio/workspace:latest
          type: string
        user:
          description: The user associated with the project
          example: daytona
          type: string
        env:
          additionalProperties:
            type: string
          description: Environment variables for the workspace
          example:
            NODE_ENV: production
          type: object
        labels:
          additionalProperties:
            type: string
          description: Labels for the workspace
          example:
            daytona.io/public: 'true'
          type: object
        public:
          description: Whether the workspace http preview is public
          example: false
          type: boolean
        target:
          description: The target environment for the workspace
          example: local
          type: string
        info:
          allOf:
            - $ref: '#/components/schemas/WorkspaceInfo'
          description: Additional information about the workspace
        cpu:
          description: The CPU quota for the workspace
          example: 2
          type: number
        gpu:
          description: The GPU quota for the workspace
          example: 0
          type: number
        memory:
          description: The memory quota for the workspace
          example: 4
          type: number
        disk:
          description: The disk quota for the workspace
          example: 10
          type: number
        state:
          allOf:
            - $ref: '#/components/schemas/WorkspaceState'
          description: The state of the workspace
          example: creating
        errorReason:
          description: The error reason of the workspace
          example: The workspace is not running
          type: string
        snapshotState:
          description: The state of the snapshot
          enum:
            - None
            - Pending
            - InProgress
            - Completed
            - Error
          example: None
          type: string
        snapshotCreatedAt:
          description: The creation timestamp of the last snapshot
          example: 2024-10-01T12:00:00Z
          type: string
        autoStopInterval:
          description: Auto-stop interval in minutes (0 means disabled)
          example: 30
          type: number
        volumes:
          description: Array of volumes attached to the workspace
          items:
            $ref: '#/components/schemas/WorkspaceVolume'
          type: array
      required:
        - env
        - id
        - labels
        - name
        - organizationId
        - public
        - target
        - user
      type: object
...
    WorkspaceLabels:
      example:
        labels:
          environment: dev
          team: backend
      properties:
        labels:
          additionalProperties:
            type: string
          description: Key-value pairs of labels
          example:
            environment: dev
            team: backend
          type: object
      required:
        - labels
      type: object
    PortPreviewUrl:
      example:
        url: https://123456-myworkspace.node.com
        token: ul67qtv-jl6wb9z5o3eii-ljqt9qed6l
      properties:
        url:
          description: Preview url
          example: https://123456-myworkspace.node.com
          type: string
        token:
          description: Access token
          example: ul67qtv-jl6wb9z5o3eii-ljqt9qed6l
          type: string
      required:
        - token
        - url
      type: object
...
    getBuildLogs_400_response:
      example:
        message: message
        statusCode: 0.8008281904610115
      properties:
        message:
          type: string
        statusCode:
          type: number
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect